Refactor: extract ConnectionDialog port-forwarding into a mixin
Move the port-forwarding rule management out of the ConnectionDialog
god-object into a new ConnectionDialogPortForwardingMixin
(sshpilot/connection_dialog_port_forwarding.py). The 10 methods
(load_port_forwarding_rules, the add/edit/delete/view handlers, the rule
editor, the per-type defaults, the listening-ports info dialog, and the
autosave no-op) are moved verbatim; self stays identical so behaviour is
unchanged. connection_dialog.py drops ~580 lines (3878 -> 3300).

get_port_checker now resolves in the new module's namespace, so its import
is removed from connection_dialog.py and the test that patches it is
repointed to the mixin module.

Adds a contract test (inheritance + per-method ownership) plus pure
per-type default-seeding coverage.
